Move code to more appropriate location
