+++ lib/open_api_spex/plug/cache.ex
@@ -21,7 +21,7 @@ defmodule OpenApiSpex.Plug.Cache do
   ```
   """
 
-  alias Plug.Conn
+  alias OpenApiSpex.OpenApi
 
   default_adapter =
     if(function_exported?(:persistent_term, :info, 0)) do
@@ -31,10 +31,9 @@ defmodule OpenApiSpex.Plug.Cache do
     end
 
   @default_adapter default_adapter
-  @missing_oas_key "Missing private.open_api_spex key in conn. Check that PutApiSpec is being called in the Conn pipeline"
 
-  @callback get(module) :: {OpenApiSpex.OpenApi.t(), map} | nil
-  @callback put(module, {OpenApiSpex.OpenApi.t(), map}) :: :ok
+  @callback get(module) :: {OpenApi.t(), map} | nil
+  @callback put(module, {OpenApi.t(), map}) :: :ok
   @callback erase(module) :: :ok
 
   @doc """
@@ -52,61 +51,6 @@ defmodule OpenApiSpex.Plug.Cache do
   def refresh do
     adapter = adapter()
     adapter.erase()
-  end
-
-  @spec get_spec_and_operation_lookup(Conn.t()) ::
-          {spec :: OpenApiSpex.OpenApi.t(),
-           operation_lookup :: %{any => OpenApiSpex.Operation.t()}}
-  def get_spec_and_operation_lookup(conn) do
-    spec_module = spec_module(conn)
-    adapter = adapter()
-    spec_and_lookup = adapter.get(spec_module)
-
-    if spec_and_lookup do
-      spec_and_lookup
-    else
-      spec_and_lookup = build_spec(spec_module)
-      adapter.put(spec_module, spec_and_lookup)
-      spec_and_lookup
-    end
-  end
-
-  @spec build_spec(module) ::
-          {OpenApiSpex.OpenApi.t(), %{String.t() => OpenApiSpex.Operation.t()}}
-  def build_spec(mod) do
-    spec = mod.spec()
-    operation_lookup = build_operation_lookup(spec)
-    {spec, operation_lookup}
-  end
-
-  @spec build_operation_lookup(OpenApiSpex.OpenApi.t()) :: %{
-          String.t() => OpenApiSpex.Operation.t()
-        }
-  defp build_operation_lookup(spec = %OpenApiSpex.OpenApi{}) do
-    spec
-    |> Map.get(:paths)
-    |> Stream.flat_map(fn {_name, item} -> Map.values(item) end)
-    |> Stream.filter(fn x -> match?(%OpenApiSpex.Operation{}, x) end)
-    |> Stream.map(fn operation -> {operation.operationId, operation} end)
-    |> Enum.into(%{})
-  end
-
-  @spec get_and_cache_controller_action(Conn.t(), String.t(), {module, atom}) ::
-          OpenApiSpex.Operation.t()
-  def get_and_cache_controller_action(conn, operation_id, {controller, action}) do
-    spec_module = spec_module(conn)
-    adapter = adapter()
-    {spec, operation_lookup} = get_spec_and_operation_lookup(conn)
-    operation = operation_lookup[operation_id]
-    operation_lookup = Map.put(operation_lookup, {controller, action}, operation)
-    adapter.put(spec_module, {spec, operation_lookup})
-    operation
-  end
-
-  @spec spec_module(Conn.t()) :: module
-  def spec_module(conn) do
-    private_data = Map.get(conn.private, :open_api_spex) || raise @missing_oas_key
-    private_data.spec_module
   end
 end
