+++ lib/open_api_spex/plug/cast.ex
@@ -60,7 +60,7 @@ defmodule OpenApiSpex.Plug.Cast do
         operation_id: operation_id,
         render_error: render_error
       }) do
-    {spec, operation_lookup} = OpenApiSpex.Plug.Cache.get_spec_and_operation_lookup(conn)
+    {spec, operation_lookup} = OpenApiSpex.Plug.PutApiSpec.get_spec_and_operation_lookup(conn)
     operation = operation_lookup[operation_id]
 
     content_type =
