+++ lib/open_api_spex/plug/put_api_spec.ex
@@ -12,7 +12,12 @@ defmodule OpenApiSpex.Plug.PutApiSpec do
 
       plug OpenApiSpex.Plug.PutApiSpec, module: MyAppWeb.ApiSpec
   """
+  alias OpenApiSpex.OpenApi
+
   @behaviour Plug
+
+  @cache OpenApiSpex.Plug.Cache.adapter()
+  @missing_oas_key "Missing private.open_api_spex key in conn. Check that PutApiSpec is being called in the Conn pipeline"
 
   @impl Plug
   def init(opts) do
@@ -29,5 +34,39 @@ defmodule OpenApiSpex.Plug.PutApiSpec do
       |> Map.put(:spec_module, spec_module)
 
     Plug.Conn.put_private(conn, :open_api_spex, private_data)
+  end
+
+  @spec get_spec_and_operation_lookup(Conn.t()) ::
+          {spec :: OpenApi.t(), operation_lookup :: %{any => Operation.t()}}
+  def get_spec_and_operation_lookup(conn) do
+    spec_module = spec_module(conn)
+    spec_and_lookup = @cache.get(spec_module)
+
+    if spec_and_lookup do
+      spec_and_lookup
+    else
+      spec = spec_module.spec()
+      operation_lookup = OpenApi.build_operation_lookup(spec)
+      spec_and_lookup = {spec, operation_lookup}
+      @cache.put(spec_module, spec_and_lookup)
+      spec_and_lookup
+    end
+  end
+
+  @spec get_and_cache_controller_action(Conn.t(), String.t(), {module, atom}) ::
+          Operation.t()
+  def get_and_cache_controller_action(conn, operation_id, {controller, action}) do
+    spec_module = spec_module(conn)
+    {spec, operation_lookup} = get_spec_and_operation_lookup(conn)
+    operation = operation_lookup[operation_id]
+    operation_lookup = Map.put(operation_lookup, {controller, action}, operation)
+    @cache.put(spec_module, {spec, operation_lookup})
+    operation
+  end
+
+  @spec spec_module(Conn.t()) :: module
+  def spec_module(conn) do
+    private_data = Map.get(conn.private, :open_api_spex) || raise @missing_oas_key
+    private_data.spec_module
   end
 end
